update script ipynb

+++ guided_diffusion/dataloader/img_deca_datasets.py
@@ -331,7 +331,7 @@ class DECADataset(Dataset):
                     each_cond_img = self.blur(each_cond_img, sigma=sigma)
                 elif 'dilate' in p:  # Dilate the mask
                     iters = int(p.split('=')[-1])
-                    each_cond_img = ndimage.binary_dilation(each_cond_img, iterations=iters)
+                    each_cond_img = ndimage.binary_dilation(each_cond_img, iterations=iters).astype(each_cond_img.dtype)
                 else: raise NotImplementedError("No preprocessing found.")
         return each_cond_img
